Python Nix package helper.
We augment a generated package with Python support if it includes a
`setup.py`.

This also involves being fairly careful about wrapper scripts and
cmake's configure_file command. Some packages include templates for
python scripts, we do not want to wrap the templates. Then, when we call
a wrapped script, we want bash to evaluate it rather than a Python
interpreter, so we remove explicit calls to PYTHON_EXECUTABLE from key
cmake scripts.